// src/interfaces/IProduct.tsx
import ICategory from "./ICategory";

export default interface IProduct {
  id: string;
  title: string;
  price: {
    currency: string;
    amount: number;
    decimals: number;
  };
  picture: string;
  condition: string;
  free_shipping: boolean;
  sold_quantity: number;
  description: string;
  categories: ICategory[];
  city_name: string;
}